CreateProductPage 컴포넌트 변경으로 인한 스타일 및 index.ts 정리
