+++ src/Components/UI/Molecules/Bar/index.ts
@@ -1,3 +1,4 @@
 export { default as MenuStatusBar } from "./MenuStatusBar";
 export { default as StoreCreateStatusBar } from "./StoreCreateStatusBar";
 export { default as StoreUpdateStatusBar } from "./StoreUpdateStatusBar";
+export { default as CreateProductStatusBar } from "./CreateProductStatusBar";
