+++ src/Components/UI/Molecules/Loading/index.ts
@@ -1,2 +1,3 @@
 export { default as Loading } from "./Loading";
 export { default as AdminLoadingAndGetUser } from "./AdminLoadingAndGetUser";
+export { default as CreateProductLoading } from "./CreateProductLoading";
