+++ src/Components/UI/Molecules/index.ts
@@ -1,8 +1,13 @@
 export {
   MenuStatusBar,
   StoreCreateStatusBar,
-  StoreUpdateStatusBar
+  StoreUpdateStatusBar,
+  CreateProductStatusBar
 } from "./Bar";
 export { Nav } from "./Nav";
-export { Loading, AdminLoadingAndGetUser } from "./Loading";
-export { ProductListItem, StoreListItem } from "./ListItem";
+export {
+  Loading,
+  AdminLoadingAndGetUser,
+  CreateProductLoading
+} from "./Loading";
+export { ProductListItem, StoreListItem, OptionFieldItem } from "./ListItem";
