+++ src/Components/UI/Organisms/Containers/index.ts
@@ -4,3 +4,5 @@ export { default as SignupFormContainer } from "./SignupFormContainer";
 export { default as UserFormContainer } from "./UserFormContainer";
 export { default as StoreFormContainer } from "./StoreFormContainer";
 export { default as LoginFormContainer } from "./LoginFormContainer";
+export { default as CreateProductFormContainer } from "./CreateProductFormContainer";
+export { default as OptionFieldContainer } from "./OptionFieldContainer";
