+++ src/Page/Admin/Product/styles.ts
@@ -1,9 +1,4 @@
-import {
-  BasicSquareButton,
-  IconButton,
-  InputDefault,
-  LabelDefault
-} from "@/Components/UI/Atoms";
+import { IconButton, InputDefault, LabelDefault } from "@/Components/UI/Atoms";
 import styled, { css } from "styled-components";
 import { Body2, Headline2, SubTitle2 } from "@/lib/styles";
 import { IAdminMenuProps } from "../interface";
@@ -94,41 +89,6 @@ export const AddimageUrl = styled(InputDefault)`
 export const AddimageUrlLabel = styled(LabelDefault)`
   cursor: pointer;
   color: ${(props) => props.theme.color.primary700};
-`;
-
-export const ButtonContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  padding: 0.8rem;
-  align-items: center;
-  justify-content: space-between;
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  color: ${(props) => props.theme.color.fontColorBlack};
-  font-size: 2rem;
-  border-top: 1px solid ${(props) => props.theme.color.gray300};
-  button {
-    cursor: pointer;
-    border: 0;
-    font-size: 2rem;
-    color: ${(props) => props.theme.color.fontColorWhite};
-    border-radius: 0.3rem;
-    line-height: 2.8rem;
-  }
-  ${({ theme }) => theme.device.mobile} {
-    justify-content: center;
-  }
-`;
-
-export const CreateButton = styled(BasicSquareButton)`
-  background-color: ${(props) => props.theme.color.primary600};
-`;
-
-export const CancelButton = styled(BasicSquareButton)`
-  color: ${(props) => props.theme.color.fontColorWhite};
-  margin-right: 0.3rem;
 `;
 
 export const ModalChildren = styled.div`
