+++ src/index.tsx
@@ -8,14 +8,14 @@ import { ReactQueryDevtools } from "react-query/devtools";
 import { theme, GlobalStyle } from "./lib/styles";
 import App from "./App";
 import { BrowserRouter } from "react-router-dom";
-import { UserContextAPI } from "./lib/state";
+import { UserContextProvider } from "./lib/state";
 
 const queryClient = new QueryClient();
 
 ReactDOM.render(
   <React.StrictMode>
     <QueryClientProvider client={queryClient}>
-      <UserContextAPI>
+      <UserContextProvider>
         <RecoilRoot>
           <BrowserRouter>
             <ThemeProvider theme={theme}>
@@ -25,7 +25,7 @@ ReactDOM.render(
             </ThemeProvider>
           </BrowserRouter>
         </RecoilRoot>
-      </UserContextAPI>
+      </UserContextProvider>
     </QueryClientProvider>
   </React.StrictMode>,
   document.getElementById("root")
